+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- fix: Capture envelope updates session state #906
+
 ## 6.1.2
 
 - fix: Clash with KSCrash functions #905